Improve graphql-import documentation
This is the first start at making the documentation more human friendly. Previously only one use-case was explained in the example. This adds a second example which first explains why you could use it instead of the first example.

With this change the documentation in the graphql-import README will also be reduced so there is less overlap between this documentation and the README, see https://github.com/prisma/graphql-import/pull/222